About & Stylesheet
about/index.html:

<aside> throws a weird issue when viewed in anything but Safari right
now. Will Fix.

=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=

style.css:

Background image for the site is changed. Not sure if this will stick
but it follows the color of my current portfolio.

Font Weight is decreased because it looked a bit too thick to read as
body copy

Chrome was being fickle about the h1 width so I tweaked it ever so
slightly. Looks almost identical.

H2 was giving the same issue with chrome. Forcing one extra pixel
helped.

H3. Same Issue.

Border-box was needed on a bunch of elements because firefox wouldn’t
show it without. Website is coded to look right only when using
border-box.